backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
python-dotenv>=1.0
openai>=1.40
pydantic>=2.6
numpy>=1.26
faiss-cpu>=1.8.0
pdfplumber>=0.11
python-docx>=0.8.11
Pillow>=10.4
PyMuPDF>=1.24
aiofiles>=23.2
pytesseract>=0.3.10
